FaceRecognizeWithPandas/requirements.txt
appdirs==1.4.4
attrs==22.1.0
black==19.3b0
Click==8.1.3
cycler==0.11.0
kiwisolver==1.4.4
matplotlib==3.6.2
nose==1.3.7
numpy==1.23.5
opencv-contrib-python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
six==1.12.0
toml==0.10.2
virtualenv==20.16.7
xlrd==2.0.1
xmltodict==0.13.0
yapf==0.32.0
mysql-connector-python==8.0.29
protobuf==4.21.9